refactor: consolidate webhook action logic into a single file and remove unused components
