--- src/action_dispatcher/webhook/client.rs
@@ -1,4 +1,0 @@
-// Re-exported from the `omnihook` crate.
-
-pub use omnihook::{OmnihookError, WebhookClient, WebhookConfig};
-
