--- src/action_dispatcher/webhook/mod.rs
@@ -1,5 +1,0 @@
-mod action;
-mod components;
-
-pub use action::WebhookAction;
-pub use components::WebhookComponents;
