--- src/action_dispatcher/webhook/payload_builder.rs
@@ -1,8 +1,0 @@
-// Re-exported from the `omnihook` crate.
-
-pub use omnihook::{
-    DiscordPayloadBuilder, GenericWebhookPayloadBuilder, SlackPayloadBuilder,
-    TelegramPayloadBuilder, WebhookPayloadBuilder,
-};
-
-
